Some fixes with mentor's advices

+++ src/main/resources/init_db.sql
@@ -3,6 +3,6 @@ USE books;
 CREATE TABLE `books` (
     `id` BIGINT NOT NULL AUTO_INCREMENT,
     `title` VARCHAR(255),
-    `price` LONG,
+    `price` DECIMAL(10,2),
     PRIMARY KEY (`id`)
 );
